+++ spec/js/i18next-touppercase.js
@@ -1,0 +1,8 @@
+export default {
+  name: 'touppercase',
+  type: 'postProcessor',
+
+  process(value, key, options, translator) {
+    return value.toUpperCase();
+  }
+}